# Remove Enum.HasFlag usage